+++ doc/src/utils/index.rst
@@ -1,0 +1,9 @@
+.. _utils:
+
+Utilities
+---------
+
+.. toctree::
+   :maxdepth: 2
+
+   fasm
