+++ doc/src/vpr/index.rst
@@ -32,5 +32,3 @@ The purpose of VPR is to make the packing, placement, and routing stages of the 
 
    file_formats
    debug_aids
-
-   fasm
