+++ utils/fasm/src/fasm.h
@@ -1,3 +1,10 @@
+// Core of the FPGA assembly (FASM) output code.
+//
+// The netlist walker implements the core walking of the netlist and routing
+// graph with the intention of emitting the complete implemented design to the
+// output FASM file.  Once output, the FASM file is expected to completely
+// describe the features required to implement the design in a hardware
+// bitstream.
 #ifndef FASM_H
 #define FASM_H
 
@@ -18,6 +25,30 @@
 
 namespace fasm {
 
+// FASM netlist visitor.
+//
+// This netlist visitor emits FASM features to the provided ostream using
+// FASM metadata tags.  See the FASM documentation in docs/utils/fasm.rst
+// for a complete list of tags.
+//
+// High level methodology:
+//  - At each top level pb_type, root tile FASM prefix is checked.
+//  - At each pb_type, the CLB prefix is build by walking from the child to
+//    the root pb_type.
+//  - All LUTs are visited and emitted via the fasm_type/fasm_lut tags.
+//  - All route through wires are emitted via the same mechanism as LUTs.
+//  - All atom's (e.g. leaf pb_types) are visited and have their parameters
+//    emitted via fasm_params tags.
+//  - All pb_type interconnect's are visited and are emitted via fasm_mux tags.
+//  - All pb_type and pb_type mode's are checked for static feature emission
+//    via the fasm_features tag.
+//
+// Once the netlist visitor is done, in finish_impl, the routing graph is
+// walked and static features are emitted via the fasm_features tag.
+//
+// Note that FASM features are incrementally written the ostream as they are
+// seen, so the order of FASM output is arbitrary based on the walk order of
+// the netlist and routing graph.
 class FasmWriterVisitor : public NetlistVisitor {
 
   public:
@@ -30,8 +61,7 @@ class FasmWriterVisitor : public NetlistVisitor {
       // clb in visit_clb_impl stands for complex logic block.
       // visit_clb_impl is called on each top-level pb_type used in the design.
       void visit_clb_impl(ClusterBlockId blk_id, const t_pb* clb) override;
-      void visit_all_impl(const t_pb_routes &top_pb_route, const t_pb* pb,
-          const t_pb_graph_node* pb_graph_node) override;
+      void visit_all_impl(const t_pb_routes &top_pb_route, const t_pb* pb) override;
       void finish_impl() override;
 
   private:
