+++ vpr/src/base/netlist_walker.cpp
@@ -14,50 +14,39 @@ void NetlistWalker::walk() {
 
         //Visit the top-level block
         visitor_.visit_clb(blk_id, pb);
+
         //Visit all the block's primitives
-        walk_blocks(pb->pb_route, pb, pb->pb_graph_node);
+        walk_blocks(pb->pb_route, pb);
     }
 
     visitor_.finish();
 }
 
-void NetlistWalker::walk_blocks(const t_pb_routes &top_pb_route, const t_pb *pb, const t_pb_graph_node *pb_graph_node) {
-    //Recursively travers this pb calling visitor_.visit_atom() or
+void NetlistWalker::walk_blocks(const t_pb_routes &top_pb_route, const t_pb* pb) {
+    // Recursively travers this pb calling visitor_.visit_atom() or
     // visitor_.visit_open() on any of its primitive pb's.
 
-    VTR_ASSERT(pb == nullptr || pb_graph_node == pb->pb_graph_node);
-    VTR_ASSERT(pb_graph_node != nullptr);
+    VTR_ASSERT(pb != nullptr);
+    VTR_ASSERT(pb->pb_graph_node != nullptr);
 
-    visitor_.visit_all(top_pb_route, pb, pb_graph_node);
-    if (pb != nullptr && pb->child_pbs == nullptr) {
+    visitor_.visit_all(top_pb_route, pb);
+    if(pb->child_pbs == nullptr) {
+        //Primitive pb
         if (pb->name != NULL) {
             visitor_.visit_atom(pb);
         } else {
             visitor_.visit_open(pb);
         }
+        return;
     }
 
-    auto pb_type = pb_graph_node->pb_type;
-
-    int mode_index = 0;
-    if (pb != nullptr) {
-        mode_index = pb->mode;
-    }
-
-    // Recurse
-    if (pb_type->num_modes > 0) {
-        const auto *mode = &pb_type->modes[mode_index];
-        for (int i = 0; i < mode->num_pb_type_children; i++) {
-            for (int j = 0; j < mode->pb_type_children[i].num_pb; j++) {
-                if ((pb != nullptr)
-                        && (pb->child_pbs != nullptr)
-                        && (pb->child_pbs[i] != nullptr)
-                        && (pb->child_pbs[i][j].pb_graph_node != nullptr)
-                        ) {
-                    walk_blocks(top_pb_route, &pb->child_pbs[i][j], pb->child_pbs[i][j].pb_graph_node);
-                } else {
-                    walk_blocks(top_pb_route, nullptr, &pb_graph_node->child_pb_graph_nodes[mode_index][i][j]);
-                }
+    //Recurse
+    const t_pb_type* pb_type = pb->pb_graph_node->pb_type;
+    if(pb_type->num_modes > 0) {
+        const t_mode *mode = &pb_type->modes[pb->mode];
+        for(int i = 0; i < mode->num_pb_type_children; i++) {
+            for(int j = 0; j < mode->pb_type_children[i].num_pb; j++) {
+                walk_blocks(top_pb_route, &pb->child_pbs[i][j]);
             }
         }
     }
@@ -83,7 +72,7 @@ void NetlistVisitor::visit_atom_impl(const t_pb* /*atom*/) {
     //noop
 }
 
-void NetlistVisitor::visit_all_impl(const t_pb_routes & /*top_pb_route*/, const t_pb* /* pb */, const t_pb_graph_node* /* pb_graph_node */) {
+void NetlistVisitor::visit_all_impl(const t_pb_routes & /*top_pb_route*/, const t_pb* /* pb */) {
     //noop
 }
