+++ vpr/src/base/netlist_writer.cpp
@@ -26,6 +26,7 @@
 #include "path_delay.h"
 #include "atom_netlist.h"
 #include "atom_netlist_utils.h"
+#include "logic_vec.h"
 
 //Overview
 //========
@@ -1896,17 +1897,6 @@ void netlist_writer(const std::string basename, std::shared_ptr<const AnalysisDe
 //
 // File-scope function implementations
 //
-
-//Output operator for vtr::LogicValue
-std::ostream& operator<<(std::ostream& os, vtr::LogicValue val) {
-    if(val == vtr::LogicValue::FALSE) os << "0";
-    else if (val == vtr::LogicValue::TRUE) os << "1";
-    else if (val == vtr::LogicValue::DONT_CARE) os << "-";
-    else if (val == vtr::LogicValue::UNKOWN) os << "x";
-    else VTR_ASSERT(false);
-    return os;
-}
-
 
 //Returns a blank string for indenting the given depth
 std::string indent(size_t depth) {
